+++ src/utils/helper.js
@@ -34,15 +34,17 @@ async function removeEmptyAccounts(indexAddress, keyringInstance, vaultState, un
   let zeroCounter = 0;
   let accountsArray = [];
   accountsArray.push({ address: indexAddress, isDeleted: false, isImported: false, label: 'Wallet 1' });
+  let labelCounter = 2;  // as an initial wallet is already created above with label 'Wallet 1'
 
   if( recoverMechanism === 'logs'){
     for(let i=0; i < logs.length; i++){
       if (logs[i].action === 'add-account'){
         const vaultState = await keyringInstance.addNewAccount(keyring[0]);
         const newAccountAddr = Web3.utils.toChecksumAddress(vaultState.keyrings[0].accounts[vaultState.keyrings[0].accounts.length - 1])
-        const label = this.createWalletLabels('all', i+2);
         if (Web3.utils.toChecksumAddress(logs[i].address) === newAccountAddr) {
+          const label = this.createWalletLabels('all', labelCounter);
           accountsArray.push({ address: newAccountAddr, isDeleted: false, isImported: false, label, isExported: false });
+          labelCounter++;
         }
         
       }